initial refactoring from refactoring branch
