+++ tools/site_and_regional/run_neon
@@ -34,12 +34,16 @@ conda activate ctsm_pylib
 
 import os
 import sys
+import argparse
 
 # -- add python/ctsm  to path
 _CTSM_PYTHON = os.path.join(
     os.path.dirname(os.path.realpath(__file__)), os.pardir, os.pardir, "python"
 )
 sys.path.insert(1, _CTSM_PYTHON)
+
+from ctsm.site_and_regional.neon_site import NeonSite
+from ctsm.site_and_regional.arg_parse import get_parser
 
 from ctsm.site_and_regional.run_neon import main
